changed color to primary color and global context for primary color
