+++ Bravotekk.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:/Users/jordan/Library/Mobile Documents/com~apple~CloudDocs/Desktop/My Folders/Work/Coding/Tekk_frontend/Bravotekk.xcodeproj">
+      location = "self:">
    </FileRef>
 </Workspace>
