+++ Tekk/ColorExtension.swift
@@ -1,0 +1,51 @@
+//
+//  ColorExtension.swift
+//  Bravotekk
+//
+//  Created by Jordan on 10/23/24.
+//
+
+import SwiftUI
+
+extension Color {
+    static let primaryColorYellow = Color("PrimaryColor")
+    
+    // Add more custom colors as needed
+    // static let secondaryColor = Color("SecondaryColor")
+}
+
+// Fallback colors for use in previews or when asset colors are unavailable
+extension Color {
+    static let fallbackPrimaryColor = Color(hex: "F6C356")
+    
+    // Add more fallback colors as needed
+    // static let fallbackSecondaryColor = Color(hex: "123456")
+}
+
+// Helper initializer for creating colors from hex strings
+extension Color {
+    init(hex: String) {
+        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
+        var int: UInt64 = 0
+        Scanner(string: hex).scanHexInt64(&int)
+        let a, r, g, b: UInt64
+        switch hex.count {
+        case 3: // RGB (12-bit)
+            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
+        case 6: // RGB (24-bit)
+            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
+        case 8: // ARGB (32-bit)
+            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
+        default:
+            (a, r, g, b) = (1, 1, 1, 0)
+        }
+
+        self.init(
+            .sRGB,
+            red: Double(r) / 255,
+            green: Double(g) / 255,
+            blue:  Double(b) / 255,
+            opacity: Double(a) / 255
+        )
+    }
+}
