+++ Tekk/Views/OnboardingView.swift
@@ -101,7 +101,7 @@ struct OnboardingView: View {
                 Text("Let's get tekky")
                     .frame(width: 325, height: 15)
                     .padding()
-                    .background(Color(hex: "947F63"))
+                    .background(Color.primaryColorYellow)
                     .foregroundColor(.white)
                     .cornerRadius(20)
                     .font(.custom("Poppins-Bold", size: 16))
