+++ Tekk/Views/QuestionnaireTwoView.swift
@@ -138,7 +138,7 @@ struct QuestionnaireTwoView: View {
                     Text("Next")
                         .frame(width: 325, height: 15)
                         .padding()
-                        .background(Color(hex: "947F63"))
+                        .background(Color.primaryColorYellow)
                         .foregroundColor(.white)
                         .cornerRadius(20)
                         .font(.custom("Poppins-Bold", size: 16))
