+++ Tekk/Views/WelcomeView.swift
@@ -129,7 +129,7 @@ struct WelcomeView: View {
                     Text("Next")
                         .frame(width: 325, height: 15)
                         .padding()
-                        .background(Color(hex: "947F63"))
+                        .background(Color.primaryColorYellow)
                         .foregroundColor(.white)
                         .cornerRadius(20)
                         .font(.custom("Poppins-Bold", size: 16))
@@ -150,11 +150,11 @@ struct WelcomeView: View {
 
 
 // MARK: - Preview
-//struct WelcomeView_Previews: PreviewProvider {
-//    @State static var showWelcome = true // Example binding variable
-//
-//    static var previews: some View {
-//        WelcomeView(showWelcome: $showWelcome)
-//            .preferredColorScheme(.dark) // Optional: Set the color scheme for the preview
-//    }
-//}
+struct WelcomeView_Previews: PreviewProvider {
+    @State static var showWelcome = true // Example binding variable
+
+    static var previews: some View {
+        WelcomeView(showWelcome: $showWelcome)
+            .preferredColorScheme(.dark) // Optional: Set the color scheme for the preview
+    }
+}
